requirements.txt
attrs==24.2.0
black==24.8.0
cads-api-client==1.3.0
cdsapi==0.7.2
certifi==2024.8.30
cftime==1.6.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.53.1
h5netcdf==1.3.0
h5py==3.11.0
idna==3.8
kiwisolver==1.4.5
matplotlib==3.9.2
multiurl==0.3.1
mypy-extensions==1.0.0
netCDF4==1.7.1.post2
numpy==2.1.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
tomli==2.0.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
xarray==2024.7.0
